Adversarial attack with FGSM on ResNet50
This approach reduces the model's performance on the input picture, but the image of a Labrador Retriever is classified as a Golden Retriever, a very similar class,  instead of a banana, the input target class. Performing a single step might be insuficient to generate adversarial noise that misclassifies the input image as the desired category.